fixed #24 Added ScopeTransaction class which implements the IDbTransaction interface.
