+++ Source/DeclarativeSql/DeclarativeSql.csproj
@@ -42,6 +42,7 @@
   <ItemGroup>
     <Compile Include="Annotations\BindParameterPrefixAttribute.cs" />
     <Compile Include="Annotations\SequenceAttribute.cs" />
+    <Compile Include="ScopeTransaction.cs" />
     <Compile Include="Helpers\CompareSelector.cs" />
     <Compile Include="Helpers\IndexedItem.cs" />
     <Compile Include="Helpers\EnumerableExtensions.cs" />
@@ -65,7 +66,7 @@
     <Compile Include="Annotations\DbConnectionAttribute.cs" />
     <Compile Include="Annotations\ProviderNameAttribute.cs" />
     <Compile Include="Mapping\TableMappingInfo.cs" />
-    <Compile Include="Helpers\TransactionScope.cs" />
+    <Compile Include="IScopeTransaction.cs" />
   </ItemGroup>
   <ItemGroup />
   <Import Project="$(MSBuildToolsPath)\Microsoft.CSharp.targets" />
